+++ demos/svelte-app/src/routes/+page.svelte
@@ -92,6 +92,7 @@
 </script>
 
 <SequentialWorkflowDesigner
+	undoStackSize={4}
 	{definition}
 	on:definitionChanged={onDefinitionChanged}
 	{steps}