+++ examples/assets/auto-select.js
@@ -1,6 +1,6 @@
 /* global document, sequentialWorkflowDesigner */
 
-function init(placeholderId, isAutoSelectDisabled) {
+function init(id, isAutoSelectDisabled) {
 	const configuration = {
 		steps: {
 			iconUrlProvider: () => './assets/icon-task.svg',
@@ -30,9 +30,28 @@ function init(placeholderId, isAutoSelectDisabled) {
 		sequence: []
 	};
 
-	const placeholder = document.getElementById(placeholderId);
-	sequentialWorkflowDesigner.Designer.create(placeholder, startDefinition, configuration);
+	const placeholder = document.getElementById(`designer${id}`);
+	const viewport = document.getElementById(`viewport${id}`);
+	const designer = sequentialWorkflowDesigner.Designer.create(placeholder, startDefinition, configuration);
+
+	function set00Viewport() {
+		designer.setViewport({
+			scale: 1,
+			position: new sequentialWorkflowDesigner.Vector(0, 0)
+		});
+	}
+
+	function reloadViewport(vp) {
+		const x = Math.round(vp.position.x);
+		const y = Math.round(vp.position.x);
+		const scale = vp.scale.toFixed(2);
+		viewport.innerText = `Viewport x: ${x}, y: ${y}, scale: ${scale}`
+	}
+
+	designer.onViewportChanged.subscribe(reloadViewport);
+	viewport.addEventListener('click', set00Viewport, false);
+	reloadViewport(designer.getViewport());
 }
 
-init('designer1', false);
-init('designer2', true);
+init('1', false);
+init('2', true);